fix(table): keep getCopyDeepLinkAction/TableAction in the published d.ts
The helper and its TableAction type were defined inline in ResponsiveDataTable
and reached the package entry only through 'export * from ./custom'.
rollup-plugin-dts drops such nested re-exports from the bundled declarations,
so 'import { getCopyDeepLinkAction, TableAction } from @sistent/sistent'
resolved to nothing in consumers despite the runtime export existing.

Re-exporting them directly from ResponsiveDataTable is not viable: it imports
the untyped @sistent/mui-datatables, which makes the dts build resolve that
module and fail (TS7016). Move both into a new leaf module TableActions (free
of untyped imports) and re-export it explicitly from src/index.tsx, the same
workaround already used for FeedbackButton, so both the value and the type land
in the published bundle.

Verified: build exits 0 and dist/index.d.ts declares both symbols.
